Code cleanup. 1. Add Function props for entry. 2. Create DxilEntrySignature to group DxilSignature. 3. Move generation of loadinput/storeoutput into HLSignatureLower.
