+++ lib/HLSL/CMakeLists.txt
@@ -34,6 +34,7 @@ add_llvm_library(LLVMHLSL
   HLOperationLower.cpp
   HLOperationLowerExtension.cpp
   HLResource.cpp
+  HLSignatureLower.cpp
   ReducibilityAnalysis.cpp
   WaveSensitivityAnalysis.cpp
