+++ lib/HLSL/DxilModule.cpp
@@ -138,9 +138,7 @@ void DxilModule::SetShaderModel(const ShaderModel *pSM) {
   m_pSM->GetDxilVersion(m_DxilMajor, m_DxilMinor);
   m_pMDHelper->SetShaderModel(m_pSM);
   DXIL::ShaderKind shaderKind = pSM->GetKind();
-  m_InputSignature.reset(new DxilSignature(shaderKind, DXIL::SignatureKind::Input));
-  m_OutputSignature.reset(new DxilSignature(shaderKind, DXIL::SignatureKind::Output));
-  m_PatchConstantSignature.reset(new DxilSignature(shaderKind, DXIL::SignatureKind::PatchConstant));
+  m_EntrySignature = llvm::make_unique<DxilEntrySignature>(shaderKind);
   m_RootSignature.reset(new RootSignatureHandle());
 }
 
@@ -984,27 +982,27 @@ void DxilModule::RemoveUnusedResources() {
 }
 
 DxilSignature &DxilModule::GetInputSignature() {
-  return *m_InputSignature;
+  return m_EntrySignature->InputSignature;
 }
 
 const DxilSignature &DxilModule::GetInputSignature() const {
-  return *m_InputSignature;
+  return m_EntrySignature->InputSignature;
 }
 
 DxilSignature &DxilModule::GetOutputSignature() {
-  return *m_OutputSignature;
+  return m_EntrySignature->OutputSignature;
 }
 
 const DxilSignature &DxilModule::GetOutputSignature() const {
-  return *m_OutputSignature;
+  return m_EntrySignature->OutputSignature;
 }
 
 DxilSignature &DxilModule::GetPatchConstantSignature() {
-  return *m_PatchConstantSignature;
+  return m_EntrySignature->PatchConstantSignature;
 }
 
 const DxilSignature &DxilModule::GetPatchConstantSignature() const {
-  return *m_PatchConstantSignature;
+  return m_EntrySignature->PatchConstantSignature;
 }
 
 const RootSignatureHandle &DxilModule::GetRootSignature() const {
@@ -1022,16 +1020,8 @@ void DxilModule::UpdateValidatorVersionMetadata() {
   m_pMDHelper->EmitValidatorVersion(m_ValMajor, m_ValMinor);
 }
 
-void DxilModule::ResetInputSignature(DxilSignature *pValue) {
-  m_InputSignature.reset(pValue);
-}
-
-void DxilModule::ResetOutputSignature(DxilSignature *pValue) {
-  m_OutputSignature.reset(pValue);
-}
-
-void DxilModule::ResetPatchConstantSignature(DxilSignature *pValue) {
-  m_PatchConstantSignature.reset(pValue);
+void DxilModule::ResetEntrySignature(DxilEntrySignature *pValue) {
+  m_EntrySignature.reset(pValue);
 }
 
 void DxilModule::ResetRootSignature(RootSignatureHandle *pValue) {
@@ -1106,9 +1096,7 @@ void DxilModule::EmitDxilMetadata() {
   m_pMDHelper->EmitValidatorVersion(m_ValMajor, m_ValMinor);
   m_pMDHelper->EmitDxilShaderModel(m_pSM);
 
-  MDTuple *pMDSignatures = m_pMDHelper->EmitDxilSignatures(*m_InputSignature, 
-                                                           *m_OutputSignature,
-                                                           *m_PatchConstantSignature);
+  MDTuple *pMDSignatures = m_pMDHelper->EmitDxilSignatures(*m_EntrySignature);
   MDTuple *pMDResources = EmitDxilResources();
   MDTuple *pMDProperties = EmitDxilShaderProperties();
   m_pMDHelper->EmitDxilTypeSystem(GetTypeSystem(), m_LLVMUsed);
@@ -1147,7 +1135,7 @@ void DxilModule::LoadDxilMetadata() {
   const ShaderModel *loadedModule;
   m_pMDHelper->LoadDxilShaderModel(loadedModule);
   SetShaderModel(loadedModule);
-  DXASSERT(m_InputSignature != nullptr, "else SetShaderModel didn't create input signature");
+  DXASSERT(m_EntrySignature != nullptr, "else SetShaderModel didn't create entry signature");
 
   const llvm::NamedMDNode *pEntries = m_pMDHelper->GetDxilEntryPoints();
   IFTBOOL(pEntries->getNumOperands() == 1, DXC_E_INCORRECT_DXIL_METADATA);
@@ -1160,8 +1148,7 @@ void DxilModule::LoadDxilMetadata() {
   SetEntryFunction(pEntryFunc);
   SetEntryFunctionName(EntryName);
 
-  m_pMDHelper->LoadDxilSignatures(*pSignatures, *m_InputSignature,
-                                  *m_OutputSignature, *m_PatchConstantSignature);
+  m_pMDHelper->LoadDxilSignatures(*pSignatures, *m_EntrySignature);
   LoadDxilResources(*pResources);
   LoadDxilShaderProperties(*pProperties);
   m_pMDHelper->LoadDxilTypeSystem(*m_pTypeSystem.get());
